[chore] Code quality baseline: ruff clean + mypy clean (strict modules)
- pyproject.toml: ignore RUF002/RUF003 (multilingual project), SIM108,
  SIM103 (style preferences), plus S108/RUF059/E702/F841 in tests only.
- Auto-fixed 59 ruff findings (F401 unused imports, I001 import order,
  C4 comprehensions, B007 unused loop vars, UP pyupgrade modernizations).
- Manually fixed remaining 29 issues: F821 undefined `Board`,
  N806 CamelCase in function scope, RUF012 mutable class default
  (added ClassVar annotation).
- ruff format: normalized 47 files.
- mypy: fixed 5 real type errors:
  * puzzle_miner: min_delta_cp default float now matches parameter type
  * config.load_settings: field_types dict value type widened to
    Union[type, tuple[type, ...]] for (int, float) pair
  * pdn._emit_tree_line call: narrowed `tree` union via assert
  * controller._is_player_piece: bool() cast on numpy.bool_ return
  * controller.get_book_moves: explicit tuple[int, int] construction

- bandit HIGH: md5() uses usedforsecurity=False (not for crypto).

Tests: 700 passed, 10 skipped, 4 xfailed (unchanged).

+++ draughts/game/puzzle_miner.py
@@ -27,8 +27,8 @@ _DEFAULT_MIN_DELTA = 4.0
 # Difficulty mapping by delta magnitude (lower-bound inclusive).
 # Ranges: 4-6 → d2, 6-10 → d3, 10+ → d4
 _DIFF_THRESHOLDS = [
-    (10.0, 4),   # delta ≥ 10 (~5 pawns) → difficulty 4
-    (6.0, 3),    # delta ≥ 6  (~3 pawns) → difficulty 3
+    (10.0, 4),  # delta ≥ 10 (~5 pawns) → difficulty 4
+    (6.0, 3),  # delta ≥ 6  (~3 pawns) → difficulty 3
 ]
 
 
@@ -55,7 +55,7 @@ def _turn_string(ply: int) -> str:
 def mine_puzzles_from_game(
     positions: list[str],
     annotations: list,  # list[MoveAnnotation] from game_analyzer
-    min_delta_cp: int = _DEFAULT_MIN_DELTA,
+    min_delta_cp: float = _DEFAULT_MIN_DELTA,
 ) -> list[dict]:
     """Extract puzzle candidates from an analyzed game.
 
@@ -159,6 +159,7 @@ def mine_puzzles_from_game(
 # ---------------------------------------------------------------------------
 # Persistence helpers (no Qt)
 # ---------------------------------------------------------------------------
+
 
 def load_mined_puzzles() -> list[dict]:
     """Load mined puzzles from the user's personal collection.
